Based on issue #1406 modified the script to run correctly
Also changed Class name according to Classes naming convention

Signed-off-by: Shirisha Ganta <shiganta@in.ibm.com>