@shirishaganta 2019-11-14
Before Patch:
JOB ID     : 15083accedf6804440053110f3c74fae5fe31937
JOB LOG    : /home/avocado-fvt-wrapper-py2/results/job-2019-11-14T11.10-15083ac/job.log
 (1/3) perf_uprobe.py:PerfUprobe.test_uprobe:  PASS (8.03 s)
 (2/3) perf_uprobe.py:PerfUprobe.test_uprobe_return:  ERROR: Command 'perf record -o /var/tmp/tmpY1XkKw -e probe_uprobe_test:doit__return -- ./uprobe_test' failed.\nstdout: ''\nstderr: "event syntax error: 'probe_uprobe_test:doit__return'\n                     \\___ unknown tracepoint\n\nError:\tFile /sys/kernel/debu... (0.84 s)
 (3/3) perf_uprobe.py:PerfUprobe.test_uprobe_variable:  PASS (1.04 s)
RESULTS    : PASS 2 | ERROR 1 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB TIME   : 10.46 s
JOB HTML   : /home/avocado-fvt-wrapper-py2/results/job-2019-11-14T11.10-15083ac/results.html

After Patch:
JOB ID     : 9521872a2ae64905ef5d0c110a8ee1b7afa92aa8
JOB LOG    : /home/avocado-fvt-wrapper-py2/results/job-2019-11-14T15.37-9521872/job.log
 (1/3) perf_uprobe.py:PerfUprobe.test_uprobe: PASS (2.72 s)
 (2/3) perf_uprobe.py:PerfUprobe.test_uprobe_return: PASS (1.13 s)
 (3/3) perf_uprobe.py:PerfUprobe.test_uprobe_variable: PASS (1.05 s)
RESULTS    : PASS 3 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB TIME   : 5.46 s
JOB HTML   : /home/avocado-fvt-wrapper-py2/results/job-2019-11-14T15.37-9521872/results.html